Removed excess spaces before newlines.
This was a CONSTANT code cleanliness issue, though it was luckily an easy fix. All it took was regex on every .cs file using the following: ` ( *?)\n` -> `\n`

Hopefully this change, in particular, can be added to the official repository, though it is unlikely considering it has been inactive for neary a decade!